feat: split chart menu from container, fix test setup due to pennant
