+++ apps/trading/components/chart-container/index.ts
@@ -1,1 +1,2 @@
-export { ChartContainer, ChartMenu } from './chart-container';
+export { ChartContainer } from './chart-container';
+export { ChartMenu } from './chart-menu';
